Задание2/requirements.txt
requests==2.32.3
pytest==8.3.4
allure-pytest==2.13.5
flake8==7.1.1
black==24.10.0
isort==5.13.2
